move refs code to references.ts
